server/requirements.txt
python-rtmidi>=1.5.8
websockets>=12.0
